Put renderTodo function in own file
