+++ styles/desktop.css
@@ -103,7 +103,7 @@ main {
   padding-right: 1rem;
 }
 
-/* WELCOME DIV STYLES */
+/* CLOCK STYLES */
 .welcomeDiv {
   height: 10rem;
   display: flex;
